tidy: Clean up integer division warnings.
The clang-tidy "integer division result used by floating point" test
pointed out a number of potential rounding issues.  Fix most of these
by converting to a float/double before the division, whether by a an
explicit cast or by changing the integer divisor to a float divisor.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-integer-division.html

+++ mythtv/libs/libmythtv/visualisations/goom/surf3d.c
@@ -32,9 +32,9 @@ grid3d *grid3d_new (int sizex, int defx, int sizez, int defz, v3d center) {
 		x = defx;
 		while (x) {
 			--x;
-			s->vertex[x+defx*y].x = (float)(x-defx/2)*sizex/defx;
+			s->vertex[x+defx*y].x = (x-defx/2.0F)*sizex/defx;
 			s->vertex[x+defx*y].y = 0;
-			s->vertex[x+defx*y].z = (float)(y-defz/2)*sizez/defz;
+			s->vertex[x+defx*y].z = (y-defz/2.0F)*sizez/defz;
 		}
 	}
 	return g;
